#31 Refactor: 컴포넌트 분리 및 useProductUpload hook으로 연결
